remove cyclic dependencies to same file
